+++ src/pages/rds.tsx
@@ -3,9 +3,10 @@ import _ from "lodash"
 import { graphql } from "gatsby"
 import { useTransition, useSpring, animated } from "react-spring"
 import styled from "styled-components"
-import { LayoutManager } from "src/components/layoutManager"
+import { LayoutManager } from "components/layoutManager"
+import { LoadingIndicator } from "components/LoadingIndicator"
 import SEO from "../components/seo"
-import { LoadingIndicator, switchVariant } from "../utils/rds"
+import { switchVariant } from "../utils/rds"
 import {
   useFetchReactionsAndSubscribeToMore,
   Reaction,
